requirements.txt
pydle~=0.9.4
pyserial~=3.4